requirements.txt
Django~=4.0.2
djoser~=2.1.0
requests~=2.27.1
djangorestframework~=3.13.1
geopy~=2.2.0
asgiref~=3.5.0
channels~=3.0.4